fix context switch

+++ arch/risc-v/src/common/up_internal.h
@@ -115,7 +115,7 @@ extern "C"
 
 extern char _heap_start[];
 extern char _heap_end[];
-EXTERN volatile uint32_t *g_current_regs;
+EXTERN volatile uintptr_t *g_current_regs;
 EXTERN uint32_t g_idle_topstack ;
 //volatile uint32_t *g_current_regs;
 //uint32_t g_idle_topstack;
